fix build errors for ante and add ante handler connection to app

+++ x/tax/genesis.go
@@ -10,7 +10,7 @@ import (
 
 type GenesisState struct {
 	Collector  sdk.AccAddress
-	Percentage int
+	Percentage int64
 }
 
 func NewGenesisState() GenesisState {
