+++ x/tax/internal/types/types.go
@@ -5,5 +5,5 @@ import sdk "github.com/cosmos/cosmos-sdk/types"
 // TaxInfo is struct for tax querying
 type TaxInfo struct {
 	Collector  sdk.AccAddress
-	Percentage int
+	Percentage int64
 }
